requirements.txt
aioredis==1.3.1
asgiref==3.3.4
async-timeout==3.0.1
attrs==21.2.0
autobahn==21.3.1
Automat==20.2.0
autopep8==1.5.7
certifi==2021.10.8
cffi==1.14.6
channels==3.0.4
channels-redis==3.3.1
charset-normalizer==2.0.7
constantly==15.1.0
cryptography==35.0.0
daphne==3.0.2
defusedxml==0.7.1
Deprecated==1.2.13
dj-database-url==0.5.0
Django==3.2.4
django-braces==1.15.0
django-ckeditor==6.1.0
django-common-helpers==0.9.2
django-cors-headers==3.8.0
django-cron==0.5.1
django-crontab==0.7.1
django-extensions==3.1.3
django-heroku==0.3.1
django-js-asset==1.2.2
django-kronos==1.0
django-oauth-toolkit==1.5.0
django-redis==5.0.0
django-rest-framework-social-oauth2==1.1.0
django-richtextfield==1.6.1
djangorestframework==3.12.4
djangorestframework-simplejwt==5.0.0
drf-social-oauth2==1.0.9
gunicorn==20.1.0
hiredis==2.0.0
hyperlink==21.0.0
idna==3.2
incremental==21.3.0
jwcrypto==1.0
msgpack==1.0.2
oauthlib==3.1.1
Pillow==8.3.1
psycopg2==2.9.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.7.0
pycparser==2.20
PyJWT==2.3.0
pyOpenSSL==21.0.0
python-crontab==2.5.1
python-dateutil==2.8.2
python-decouple==3.4
python3-openid==3.2.0
pytz==2021.1
redis==3.5.3
requests==2.26.0
requests-oauthlib==1.3.0
service-identity==21.1.0
six==1.16.0
social-auth-app-django==5.0.0
social-auth-core==4.1.0
sqlparse==0.4.1
toml==0.10.2
Twisted==21.7.0
twisted-iocpsupport==1.0.2
txaio==21.2.1
typing-extensions==3.10.0.0
urllib3==1.26.7
whitenoise==5.3.0
wrapt==1.13.3
zope.interface==5.4.0
